+++ database/migrations/2019_05_05_175036_create_votes_table.php
@@ -13,17 +13,17 @@ class CreateVotesTable extends Migration
      */
     public function up()
     {
-        Schema::create('votes', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->integer('institution_id')->default(0);
-            $table->integer('job_id')->default(0);
-            $table->integer('program_id')->default(0);
-            $table->integer('profile_id')->default(0);
-            $table->unsignedInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('rank');
-            $table->timestamps();
-        });
+      Schema::create('votes', function (Blueprint $table) {
+        $table->bigIncrements('id');
+        $table->integer('institution_id')->default(0);
+        $table->integer('job_id')->default(0);
+        $table->integer('program_id')->default(0);
+        $table->integer('profile_id')->default(0);
+        $table->unsignedBigInteger('user_id');
+        $table->foreign('user_id')->references('id')->on('users');
+        $table->integer('rank');
+        $table->timestamps();
+      });
     }
 
     /**
